refactor(decode): narrow coefficient sign storage to i8 (#1052)
The width audit found the quantizer-sign grid holding only -1/0/1 yet
stored as i32: the padded per-block scratch, its thread-local recycler
pool, and the shared all-zero fallback were all four bytes per cell.
Store signs as i8: the 36x36 padded grid shrinks 5,184 -> 1,296 bytes
per active FSC block, per-take zeroing traffic drops to a quarter, and
the static fallback shrinks 5.1 KiB -> 1.3 KiB. Context summation
widens each sign to i32 at the single read site.

No CPU-speed claim: instruction counts on the four measured fixtures
move within layout noise (-0.2% to +1%, both directions) because none
of them is FSC-dense enough to allocate sign grids on the hot path.

Negative result, recorded and not shipped: narrowing CoeffScanEntry
(four usize for values bounded by the 32x32 transform) to u16/u8
measured +2.1% instructions on the coefficient-heavy fixture — the
per-accessor widening in the hot loop outweighs the footprint saving
at these block sizes.

Feature: INFRA-DECODE-SERIAL-HOT-PATHS

## Correctness

- raw outputs byte-identical to main on four fixtures
- sign values are AV2 sign flags (§ 8.3.2 coeff sign, -1/0/+1 by
  construction in quant_sign_value); i8 is exact
- full suite green: fmt, clippy -D warnings, 30 workspace suites,
  decoder-fixtures verify, xtask checks (run individually; this
  container's /dev/null device node is broken, see #1049)

+++ crates/splot-decode/src/bitstream/tile_payload/cdf/coeff_context.rs
@@ -271,7 +271,7 @@ pub(crate) const fn dc_sign_ctx(
 }
 
 pub(crate) const fn idtx_sign_ctx(
-    quant_sign: &[i32],
+    quant_sign: &[i8],
     level: &[u8],
     row: usize,
     col: usize,
@@ -281,19 +281,19 @@ pub(crate) const fn idtx_sign_ctx(
     if col > 0 {
         let idx = row.saturating_mul(stride).saturating_add(col - 1);
         if idx < quant_sign.len() {
-            signc += quant_sign[idx];
+            signc += quant_sign[idx] as i32;
         }
     }
     if row > 0 {
         let idx = (row - 1).saturating_mul(stride).saturating_add(col);
         if idx < quant_sign.len() {
-            signc += quant_sign[idx];
+            signc += quant_sign[idx] as i32;
         }
     }
     if col > 0 && row > 0 {
         let idx = (row - 1).saturating_mul(stride).saturating_add(col - 1);
         if idx < quant_sign.len() {
-            signc += quant_sign[idx];
+            signc += quant_sign[idx] as i32;
         }
     }
     let mut ctx: usize = if signc > 2 {
